Refactor/internal tools (#932)
* refactor: audio array

* feat: audio bus

* chore: added docs for both utils

* ci: lint

* refactor: audio array and fft

* chore: requested changes

* refactor: next part for audio bus and audio array

* refactor: wip

* fix: a few fixes

* fix: lint

* fix: fixed tests

* fix: todos

* refactor: nits

* fix: nitpick

* refactor: added copyWithin - memmove to audio array

* fix: memmove

* chore: updated custom processor template

* refactor: optimized interleaveTo function

* refactor: renaming

* refactor: renaming

* refactor: renaming

* refactor: doxygen for AudioBuffer

* refactor: deinterleave audio data using AudioBuffer method

* refactor: broader support for circular data structures operations and deinterleaving

* chore: requuested changes from code review

* fix: nits

* ci: lint

* chore: requested changes from cr

* ci: lint

* refactor: get channel by type

---------

+++ packages/react-native-audio-api/common/cpp/audioapi/HostObjects/AudioNodeHostObject.cpp
@@ -36,7 +36,7 @@ JSI_PROPERTY_GETTER_IMPL(AudioNodeHostObject, numberOfOutputs) {
 }
 
 JSI_PROPERTY_GETTER_IMPL(AudioNodeHostObject, channelCount) {
-  return {node_->getChannelCount()};
+  return {static_cast<int>(node_->getChannelCount())};
 }
 
 JSI_PROPERTY_GETTER_IMPL(AudioNodeHostObject, channelCountMode) {
